+++ src/xopr/xarray_accessor/style.css
